Fix clone() and with_...() method for AsyncClient
